% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBP.R
\name{hmm_LBP}
\alias{hmm_LBP}
\title{hmm_LBP}
\usage{
hmm_LBP(hmm)
}
\arguments{
\item{hmm}{The name of a profile hidden markov model file.}
}
\value{
A vector of length 256.
}
\description{
This feature uses local binary pattern with a neighborhood of radius 1 and 8 sample points to extract
features from the HMM. A 256 bin histogram is extracted as a 256 length feature vector.
}
\examples{
h<- hmm_LBP(system.file("extdata", "1DLHA2-7", package="protHMM"))

}
\references{
Li, Y., Li, L., Wang, L., Yu, C., Wang, Z., & You, Z. (2019).
An Ensemble Classifier to Predict Protein–Protein Interactions by Combining PSSM-based Evolutionary Information with Local Binary Pattern Model.
International Journal of Molecular Sciences, 20(14), 3511.
}
