% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{hmm_read}
\alias{hmm_read}
\title{hmm_read}
\usage{
hmm_read(hmm)
}
\arguments{
\item{hmm}{The name of a profile hidden markov model file.}
}
\value{
A 20 x L matrix, in which L is the sequence length.
}
\description{
Reads in the amino acid emission frequency columns of a profile hidden markov model matrix
and converts each position to frequencies.
}
\examples{
h<- hmm_read(system.file("extdata", "1DLHA2-7", package="protHMM"))
}
