% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AC.R
\name{hmm_ac}
\alias{hmm_ac}
\title{hmm_ac}
\usage{
hmm_ac(hmm, lg = 4)
}
\arguments{
\item{hmm}{path of a HMM output file}

\item{lg}{lag number}
}
\value{
A vector of length 20 * the lag value; by default this is a vector of length 80.
}
\description{
Calculated auto-covariance for profile hidden markov model outputs from HHblits
}
